backend/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.5.0
django-dotenv==1.4.2
djangorestframework==3.15.2
pillow==11.0.0
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
